requirements.txt
datasets
pillow
brainflow
torch
pandas
numpy
matplotlib
seaborn
PySide
scipy